# Optimized, parallelized Blur effect